UWP Mesa working

+++ Src/CPU/Z80/Z80.cpp
@@ -2479,7 +2479,7 @@ int CZ80::Run(int numCycles)
     case 0x44:      /* NEG */
       cycles -= cycleTables[2][0x44];
       temp = hreg(AF);
-      AF = (-(AF & 0xff00) & 0xff00);
+      AF = ((unsigned int)-(int)(AF & 0xff00) & 0xff00);
       AF |= ((AF >> 8) & 0xa8) | (((AF & 0xff00) == 0) << 6) |
         (((temp & 0x0f) != 0) << 4) | ((temp == 0x80) << 2) |
         2 | (temp != 0);
